refactor utilities to testibng, add CLI tests
